requirements.txt
numpy>=1.16.2
networkx>=2.2
gensim==3.4.0
mlflow
google-cloud-storage
cloudml-hypertune
gcsfs
